notification to manager in progress

+++ LeaveAppManagement.dataAccess/Dto/LeaveBalanceDto.cs
@@ -6,8 +6,8 @@ namespace LeaveAppManagement.dataAccess.Dto
     {
         public int EmployeeId { get; set; }
         public string EmployeeName { get; set; } = string.Empty;
-        public int TotaLeaveAvailable { get; set; }
-        public int TotalLeaveUsed { get; set; }
-        public int TotalCurrentLeave { get; set; }
+        public double TotaLeaveAvailable { get; set; }
+        public double TotalLeaveUsed { get; set; }
+        public double TotalCurrentLeave { get; set; }
     }
 }
