+++ LeaveAppManagement.dataAccess/Dto/UsersDto.cs
@@ -11,7 +11,7 @@ namespace LeaveAppManagement.dataAccess.Dto
         public string Password { get; set; } = string.Empty;
         public string PhoneNumber { get; set; } = string.Empty;
         public string Job { get; set; } = string.Empty;
-        public int TotaLeaveAvailable { get; set; }
+        public double TotaLeaveAvailable { get; set; }
         public DateTime HireDate { get; set; }
         public int RoleId { get; set; }
         public string? RoleName { get; set; } = null;
