+++ LeaveAppManagement.dataAccess/Models/LeaveBalance.cs
@@ -7,7 +7,7 @@ namespace LeaveAppManagement.dataAccess.Models
         public int Id { get; set; }
         public int Years { get; set; }
         public int EmployeeId { get; set; }
-        public int TotalCurrentLeave { get; set; }
+        public double TotalCurrentLeave { get; set; }
         public Employee? Employee { get; set; }
         public int LeaveRequestId { get; set; }
         public virtual LeaveRequest? LeaveRequest { get; set; }
